put component's together, and works :)!
