--- src/App.css
@@ -1,21 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-}
-
-.fullPage {
-  height: 100vh;
-  width: 100%;
-}
-
-.main {
-  width: 100vw;
-  height: 100vh;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.bg1 {
-  background-color: #ff0;
-}
